chore(dev): Wrap `vector-config` in `vector-lib` as `configurable` (#18944)
* chore(dev): Wrap `vector-config` in `vector-lib` as `configurable`

There already exists a `vector_lib::config` export so I figured adding
`vector-config` to that module would make things too confusing. Since most of
`vector-config` has to do with the `configurable` terminology, that seemed a
natural fit for this export.

Note that, unlike the previous wrappers, we are unable to get rid of the import
of `vector-config` into the base package. This is because the macros in
`vector-config-macros` explicitly assume the use of `vector-config` in their
coding.

* Add package name hack to avoid needing to import `vector-config` in the root

* Fixup stray FIXME

* Drop the imports of `vector-config-*` into `vector-lib`