feat(new source): Add Prometheus Pushgateway source (#18143)
* Start work on Pushgateway source

I have no idea what I'm doing.

* Fix typo

* Add acknowledgements config

I don't know what this does yet, but it seems like something I can
figure out later. Also it's deprecated so maybe this is fine.

* Start implementing source

* Minimal (untested) implementation for HttpSource

* Remove unused imports

* Fix compile errors

* Properly register Pushgateway config

* Actually parse Prometheus metrics

* Leave myself some TODOs

* Start parsing path to produce labels

* "Split grouping key into pairs"

* Slight refactor

* Add comments, avoid pointless allocation

* Fix comment

* Fix compile failures

* Decode base64 labels

Feels like there has to be a nicer way to handle the errors, but this
will do for a start.

* Do that label decoding refactor

* Use labels from path when generating metric events

Currently replaces all tags on the metric event, should only replace the
ones that were passed in the URL.

* Add path labels without losing all others

* Extract function

* Implement overrides for all metric types

* Remove TODOs

The path one won't work as the Prometheus client libraries always use a
fixed path prefix. The other one is done :D

* Add metric aggregation

* Add Cargo.lock update

* Use more general type parameter

The dance to and from MetricTags felt really pointless. Let's not do it.

* Update comments

* Return HTTP error when we fail to parse body

* Remove println from dev

* Add "pushgateway" to allowed words

* Remove redundant into_iter/collect

* Remove pointless cloning

* Improve type signature

* Remove pointless assignment

* Switch from 'from' to 'into'

* cargo fmt

* Fix clippy

* I was wrong, clippy was right

* Fix duplicate import in Cargo.toml

* Catch up with upstream method signature change

* Always treat summaries as absolute

Aggregating summary values results in nonsense, so let's not do that.

If we receive a summary for a labelset we already have a value for, we
should replace the existing value.

* Generate docs

* Add compile feature: sources-prometheus-pushgateway

* Fix Prometheus feature

* Start adding pushgateway tests

* Test base64 parsing and fix bug

This behaviour was wrong and the test caught that.

* Enforce that job has a value

* Ensure path has valid prefix

* Improve tests and fix another bug

* Add clarifying comment

* Sketch parser tests

* Delete old comment

* Add TODO

* WIP for tomorrow

* more tests

* match other tests

* Refactor test helpers

Feels better to collapse these down to one helper and avoid the
confusing shadowing of the real methods.

* More tests for label overrides

* Add tests for metric aggregation

* Add clarification about trailing slashes in path

* Add test for config generation

* Add full source-level tests

* Fix spelling

* Fix Clippy and remove completed TODOs

* Fix fmt

* fmt

* Clarify grouping key error messages

* Update pushgateway.rs to match upstream refactor

A bunch of modules been renamed upstream, so let's match the new names.

* Update module names in Prometheus parser tests

* Allow base64 padding to be omitted in Prometheus Pushgateway source

The Prometheus Pushgateway has a fairly permissive base64 implementation
that allows padding to be missing. We need to fake that by adding in any
missing padding before we pass the value to the base64 decoder.

This is documented, as examples in their README don't use padding:

https://github.com/prometheus/pushgateway/blob/ec7afda4eef288bd9b9c43d063e4df54c8961272/README.md#url

* Don't spellcheck base64 string

* Hopefully correctly ignore spelling?

* Add missing newline

* Add module documentation to pushgateway.rs

* Fix documentation about types of metric that can be aggregated

This statement was about a previous version of the code which turned out
to be wrong (it was aggregating summaries).

* Add documentation (CUE) for Prometheus Pushgateway

* Add documentation (Markdown) for Prometheus Pushgateway

* cargo fmt

* Regenerate component docs

* Update src/sources/prometheus/pushgateway.rs

Co-authored-by: Stephen Wakely <stephen@lisp.space>

* Add `prometheus_pushgateway` link to URL list for docs

* Sort `prometheus_*` URLs in urls.cue

* Run `cue fmt`

* Regenerate docs

* retrigger CI

* Add changelog entry

* Add 'Sinjo' to list of allowed words for spell check

This allows my username to appear in markdown files.

* Remove deprecated internal metrics from docs

These have been removed in `0.35.0` which has landed on `master`.

---------

Co-authored-by: Stephen Wakely <stephen@lisp.space>
Co-authored-by: neuronull <neuronull@pm.me>