feat(sources)!: add TCP mode to DNSTAP source (#19892)
* feat(sources): add TCP mode to DNSTAP source

This adds TCP mode to DNSTAP source, while leaving the current one ("unix") as the default. This
required a minor hack (https://stackoverflow.com/questions/61216723/how-can-i-deserialize-an-enum-with-an-optional-internal-tag/61219284#61219284)
in serde deserialization. Unix is the default when `cfg!(unix)` is true, otherwise TCP is the
default on port 9000. This can be changed to never default to TCP without explicitly `mode` to TCP.

This is implemented as a different framestream source, reusing some parts of TCP source code and
some parts of framestream unix source.

Fixes: #19735

* Add tests for TCP framestream

* Add documentation for TCP mode

* Add missing configuration options for DNSTAP TCP mode

* Fix dnstap unix integration test

* Support tls client metadata for dnstap TCP source

* Add changelog entry

* Fix fomatting in DNSTAP TCP address docs

Co-authored-by: Ursula Chen <58821586+urseberry@users.noreply.github.com>

* Remove @ from authors entry in changelog

* Fix warnings in framestream.rs

* Run generate-component-docs

* Fix broken generated logs

Since DNSTAP config is a bit different to other configs since it defaults to `unix` mode (so it had
to be untagged for that to work), it does not generate proper docs. This adds overrides to properly
flag different fields for `unix` or `tcp` mode.

* Fix `TcpConfig` serde tag attribute condition

* Add origin permit config options to DNSTAP in TCP mode

* Set `permit_origin` option to be relevant only in tcp mode

* Add test case for permit_origin

* Generate component docs

* Add `ipallowlist` to spellcheck allowed words

* Reduce duplication in DNSTAP source

* Make `permit_origin` optional

* Update component docs for dnstap

* Make `mode` required for `dnstap` source

* Update changelog entry

* Add must_use to `with_allowlist`

* Fix issues in tests

* Fix clippy warnings

* Enable std for `ipnet` for tests

* Add ipnet dependency to vector-core

Co-authored-by: Stephen Wakely <stephen@lisp.space>

* Make `tcp` depend on `ipnet` and `dnstap` depend on `tcp`

---------

Co-authored-by: Ursula Chen <58821586+urseberry@users.noreply.github.com>
Co-authored-by: Stephen Wakely <stephen@lisp.space>